requirements.txt
opencv-python
numpy
Pillow
pyOpenSSL
